Make autologin a little more dynamic, and simplify.
# Added multi os functionality as kcpassword files have changed across
major versions. Add your payload in the format as follows;
# kcpassword.10.6.8 or kcpassword.10.7.4 or kcpassword.10.7.5 etc etc

+++ PAYLOAD/AppleAutoLoginUser.sh
@@ -2,8 +2,7 @@
 # AppleAutoLoginUser : Payload required. Auto login a specified user.
 # chris.gerke@gmail.com
 
-CONFIG=$(sudo defaults read "$1/Contents/Resources/config" autoLoginUser)
-KCPASSWORD="$1/Contents/Resources/PAYLOAD/kcpassword"
+SHORTNAME=""
 
 # // fix
 ROOT="$3"; if [ -z "${ROOT}" ] || [ "${ROOT}" = "/" ]; then ROOT=""; fi
@@ -11,20 +10,13 @@ ROOT="$3"; if [ -z "${ROOT}" ] || [ "${ROOT}" = "/" ]; then ROOT=""; fi
 # Multi OS
 TARGET_OS=$(sudo defaults read "${ROOT}/System/Library/CoreServices/SystemVersion" ProductVersion)
 
-if [ -z "${CONFIG}" ] || [ "${CONFIG}" != "" ]; then
- # Payload
- if [ -r "${KCPASSWORD}" ]; then
-  # kcpassword
-  sudo cp -f "${KCPASSWORD}" "${ROOT}/etc/kcpassword"
-  sudo chown root:wheel "${ROOT}/etc/kcpassword"
-  sudo chmod 600 "${ROOT}/etc/kcpassword"
-  sudo rm -Rf "${ROOT}/etc/kcpassword.disabled"
-  # com.apple.loginwindow.plist
-  sudo defaults write "${ROOT}/Library/Preferences/com.apple.loginwindow" autoLoginUser -string "${CONFIG}"
-  sudo rm -Rf "${ROOT}/Library/Preferences/com.apple.loginwindow.plist.lockfile"
-  sudo chown root:wheel "${ROOT}/Library/Preferences/com.apple.loginwindow.plist"
-  sudo chmod 644 "${ROOT}/Library/Preferences/com.apple.loginwindow.plist"
- fi
-fi
+# Payload
+[ -e "$1/Contents/Resources/PAYLOAD/kcpassword.${TARGET_OS}" ] && sudo cp -f "$1/Contents/Resources/PAYLOAD/kcpassword" "${ROOT}/etc/kcpassword"; echo "Payload"
+[ -e "${ROOT}/etc/kcpassword" ] && sudo chown root:wheel "${ROOT}/etc/kcpassword"
+[ -e "${ROOT}/etc/kcpassword" ] && sudo chmod 600 "${ROOT}/etc/kcpassword"
+[ -e "${ROOT}/etc/kcpassword.disabled" ] && sudo rm -Rf "${ROOT}/etc/kcpassword.disabled"
+ 
+# com.apple.loginwindow.plist
+[ -e "${ROOT}/etc/kcpassword" ] || [ -z "${SHORTNAME}" ] || [ "${SHORTNAME}" != "" ] && sudo defaults write "${ROOT}/Library/Preferences/com.apple.loginwindow" autoLoginUser -string "${SHORTNAME}"; sudo rm -Rf "${ROOT}/Library/Preferences/com.apple.loginwindow.plist.lockfile"; sudo chown root:wheel "${ROOT}/Library/Preferences/com.apple.loginwindow.plist"; sudo chmod 644 "${ROOT}/Library/Preferences/com.apple.loginwindow.plist"; echo "autoLoginUser"
 
 exit 0
